fix random time y esconderlo de maincamera

+++ ufide-videojuegos-1-hogans-alley/Assets/Scripts/level1.cs
@@ -24,7 +24,7 @@ public class level1 : MonoBehaviour {
     /*Guys*/ int totalBadGuys = 0; bool[] isBadGuy = {false, false, false}, isShot; string[] guy; 
     /*Score*/ bool[] wasFastShot = {false, false, false};
     /*Sharpshooter*/ public int sharpshooterLevel = 15, superSharpshooterLevel = 30;
-    /*Tiempo*/ public float maxTime = 2.50f, minMaxTime = 1.30f, minTime = 0.60f, extraTime = 3f, randomTime = 0.00f;
+    /*Tiempo*/ float maxTime = 2.50f, minMaxTime = 1.30f, minTime = 0.60f, extraTime = 3f, randomTime = 0.00f;
     /*x round y miss digits*/ float r1x = 0.35f, r2x = -0.35f; float m1x = 1.1f, m2x = 0.7f;
     /*x score digits*/ float s3x = -3.55f, s4x = -3.95f, s5x = -4.35f, s6x = -4.75f;
     /*x time digits*/ float time1x = 0.4f, time2x = -0.4f;
@@ -166,7 +166,7 @@ public class level1 : MonoBehaviour {
 
     /*Listo*/ public int generateRandomTime() {
         timetime = UnityEngine.Random.Range( Convert.ToInt32((maxTime*10)-10), Convert.ToInt32(maxTime*10) );
-        randomTime = timetime/10;
+        randomTime = (float) timetime/10;
         if (randomTime < minTime ) { randomTime = minTime; timetime = Convert.ToInt32(randomTime)*10; }
         if (maxTime > minMaxTime) { maxTime = maxTime - 0.05f; }
         return timetime;
